+++ packages/playground/src/bots/GenesisCoinbaseBot/__providers.ts
